requirements_ekos.txt
accelerate==0.27.2
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
archspec @ file:///croot/archspec_1697725767277/work
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
boltons @ file:///work/ci_py311/boltons_1677685195580/work
braceexpand==0.1.7
Brotli @ file:///work/ci_py311/brotli-split_1676830125088/work
cachetools==5.3.2
catalogue==2.0.10
certifi @ file:///croot/certifi_1707229174982/work/certifi
cffi @ file:///croot/cffi_1700254295673/work
cfgv==3.4.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cloudpathlib==0.16.0
comm==0.2.2
conda @ file:///croot/conda_1707839125844/work
conda-content-trust @ file:///croot/conda-content-trust_1693490622020/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1702997573971/work/src
conda-package-handling @ file:///croot/conda-package-handling_1690999929514/work
conda_package_streaming @ file:///croot/conda-package-streaming_1690987966409/work
confection==0.1.4
ConfigArgParse==1.7
contexttimer==0.3.3
contourpy==1.2.0
cryptography @ file:///croot/cryptography_1702070282333/work
cycler==0.12.1
cymem==2.0.8
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==2.17.0
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
diffusers==0.16.0
dill==0.3.8
distlib==0.3.8
distro @ file:///croot/distro_1701455004953/work
docker-pycreds==0.4.0
easydict==1.13
einops==0.7.0
executing==2.0.1
fairscale==0.4.4
fastjsonschema==2.19.1
filelock==3.13.1
fire==0.5.0
Flask==3.0.3
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
gdown==5.1.0
gitdb==4.0.11
GitPython==3.1.42
h5py==3.10.0
huggingface-hub==0.20.3
identify==2.5.35
idna @ file:///work/ci_py311/idna_1676822698822/work
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
iopath==0.1.10
ipdb==0.13.13
ipython==8.21.0
ipywidgets==8.1.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kaggle==1.6.6
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
libmambapy @ file:///croot/mamba-split_1698782620632/work/libmambapy
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
menuinst @ file:///croot/menuinst_1702390294373/work
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
open3d==0.18.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
orjson==3.9.15
packaging @ file:///croot/packaging_1693575174725/work
pandas==2.2.0
parso==0.8.3
peft==0.8.2
pexpect==4.9.0
pillow==10.2.0
platformdirs @ file:///croot/platformdirs_1692205439124/work
plotly==5.19.0
pluggy @ file:///work/ci_py311/pluggy_1676822818071/work
portalocker==2.8.2
pre-commit==3.6.2
preshed==3.0.9
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycocoevalcap==1.2
pycocotools==2.0.7
pycosat @ file:///croot/pycosat_1696536503704/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
pyinstrument==4.6.2
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.1
pyquaternion==0.9.9
PySocks @ file:///work/ci_py311/pysocks_1676822712504/work
python-dateutil==2.8.2
python-magic==0.4.27
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests @ file:///croot/requests_1690400202158/work
retrying==1.3.4
rich==13.7.0
rpds-py==0.18.0
ruamel.yaml @ file:///work/ci_py311/ruamel.yaml_1676838772170/work
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
sentencepiece==0.1.99
sentry-sdk==1.40.5
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
streamlit==1.31.1
supervision==0.18.0
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
thinc==8.2.3
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm @ file:///croot/tqdm_1679561862951/work
traitlets==5.14.1
transformers==4.31.0
triton==2.2.0
truststore @ file:///croot/truststore_1695244293384/work
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3 @ file:///croot/urllib3_1698257533958/work
validators==0.22.0
virtualenv==20.25.0
wandb==0.16.3
wasabi==1.1.2
watchdog==4.0.0
wcwidth==0.2.13
weasel==0.3.4
webdataset==0.2.86
webencodings==0.5.1
Werkzeug==3.0.3
widgetsnbextension==4.0.10
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
zstandard @ file:///work/ci_py311_2/zstandard_1679339489613/work
